+++ ouds_core/lib/l10n/gen/ouds_localizations.dart
@@ -287,6 +287,30 @@ abstract class OudsLocalizations {
   /// **'TextField'**
   String get core_text_input_input_a11y;
 
+  /// No description provided for @core_password_input_hidden_a11y.
+  ///
+  /// In en, this message translates to:
+  /// **'Password hidden'**
+  String get core_password_input_hidden_a11y;
+
+  /// No description provided for @core_password_input_visible_a11y.
+  ///
+  /// In en, this message translates to:
+  /// **'Password visible'**
+  String get core_password_input_visible_a11y;
+
+  /// No description provided for @core_password_input_hint_show_a11y.
+  ///
+  /// In en, this message translates to:
+  /// **'Double tap to show password'**
+  String get core_password_input_hint_show_a11y;
+
+  /// No description provided for @core_password_input_hint_hide_a11y.
+  ///
+  /// In en, this message translates to:
+  /// **'Double tap to hide password'**
+  String get core_password_input_hint_hide_a11y;
+
   /// No description provided for @core_pin_code_input_digit_code_label_a11y.
   ///
   /// In en, this message translates to: